+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 ## Unreleased
 
+## v2.1.0
+
+* [Reorder issues](./_FEATUREBOOK/featurebook/ReorderTad.md)
+* [Organize custom queries](./_FEATUREBOOK/featurebook/OrganizeCustomQueriesTad.md)
+
 ## v2.0.0
 
 * [Quick edit field](./_FEATUREBOOK/featurebook/QuickEditFieldTad.md), migrated from an internal plugin project. W/ some improvements: